pinenote-arch: integrate ~phantomas/pinenote-config-sway
Setup pinenote-config-sway using default path.

Regarding the fragment configuration:
Since this repo and the config had slightly different setup for the menu
and lisgd, the $menu and $gestures_service are set to the scripts that
were previously in use.

The launch_lisgd script has been updated with two new gestures to show &
hide the sidebar, and the refresh-screen gesture now using the dbus
object from pinenote-service.

The rotation service in use is still rot8, but I'm using the one wrapper
shipped in pinenote-config-sway, since the waybar config expect it to
behave in a specific way.

The default pre-defined render hints are ported from
sway_dbus_integration.py.

The following changes were made since v1:
- Add some documentation comments to explain what's being done, and
  where some variable are coming from.
- Improved consistency in variable names. The $pn_ variable are all
  coming from the config fragment, while the one without prefix are
  meant to be overridden in the config.
- Add a link back to my repository since the readme is now up to date
- Removed commented out code
- Improved quoting consistency
- Bump pinenote-arch package version